+++ fern/versions/latest/pages/model-server/model-call-capture.mdx
@@ -55,7 +55,7 @@ Concurrent collection jobs with overlapping indices must use separate capture di
 ## Read captured calls
 
 ```python
-from nemo_gym.model_call_capture import (
+from nemo_gym.base_responses_api_model import (
     CaptureStore,
     aggregate_model_call_metrics,
     read_model_call_records,
